requirements.txt
lxml
requests
python-dotenv
networkx
pyvis
matplotlib